requirements.txt
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
PyYAML>=6.0
tqdm>=4.66
joblib>=1.3
scikit-image>=0.22
tqdm-joblib>=0.0.4
